Complete modular refactoring: app.py reduced from 2500+ to 94 lines
This is a comprehensive refactoring that transforms the monolithic
app.py into a well-organized, modular codebase.

STRUCTURE:
==========

vectorytones/
├── audio/ - Audio generation & processing utilities
├── datasets/ - Dataset management
│   ├── config.py - DEMO_DATASETS configuration
│   ├── downloader.py (157 lines) - Download functions
│   └── loader.py (865 lines) - Loading functions
├── models/ - ML models, embeddings & training
│   ├── loader.py - Model initialization
│   ├── embeddings.py - Embedding functions
│   └── training.py - ML training
├── routes/ - Flask blueprints (852 lines total)
│   ├── main.py (21 lines) - / and /favicon.ico
│   ├── clips.py (150 lines) - Clip routes
│   ├── sorting.py (447 lines) - Sorting/voting routes
│   └── datasets.py (234 lines) - Dataset routes
└── utils/ - Progress tracking & state management

config.py (35 lines) - All configuration constants
app.py (94 lines) - Minimal initialization only

CHANGES:
========

app.py:
- Reduced from 2500+ lines to 94 lines (96% reduction)
- Now only handles initialization and blueprint registration
- All 27 routes extracted to Flask blueprints
- All 46 functions extracted to proper modules

New Flask Blueprints:
- main_bp: Static file serving
- clips_bp: All /api/clips/* endpoints
- sorting_bp: Sorting, voting, and ML routes
- datasets_bp: Dataset management endpoints

Dataset Functions Extracted:
- download_esc50(), download_cifar10(), download_ucf101_subset()
- download_20newsgroups(), download_file_with_progress()
- load_dataset_from_folder(), load_demo_dataset()
- All metadata loading functions

BENEFITS:
=========

✅ 96% reduction in main file size
✅ Proper separation of concerns
✅ Highly testable modular design
✅ Easy to maintain and extend
✅ All functionality preserved
✅ Ready for CI/CD and testing frameworks
